add app.ts
